Moved CreateGift type to shard/types.ts and fixed all of the imports
